Fix amici extension import logo for Windows (#1084)
and import module on travis

fixup

cd

dbg info

dbg

path

path

path

fixup

Fix msvc compiler warnings (#1114)

path

Fix msvc compiler warnings (#1114)

Fix msvc compiler warnings (#1114)

path?

ls

Fix msvc compiler warnings (#1114)

ls

path?

-ls

path

path?

path?

..

walk

walk

..

d

+++ src/newton_solver.cpp
@@ -107,7 +107,7 @@ void NewtonSolver::computeNewtonSensis(AmiVectorArray &sx) {
                 auto data_ptr = model->dxdotdp_explicit.data();
                 for (sunindextype iCol = col[model->plist(ip)];
                      iCol < col[model->plist(ip) + 1]; ++iCol)
-                    sx.at(row[iCol], ip) -= data_ptr[iCol];
+                    sx.at(static_cast<int>(row[iCol]), ip) -= data_ptr[iCol];
             }
 
             // copy implicit version
@@ -117,7 +117,7 @@ void NewtonSolver::computeNewtonSensis(AmiVectorArray &sx) {
                 auto data_ptr = model->dxdotdp_implicit.data();
                 for (sunindextype iCol = col[model->plist(ip)];
                      iCol < col[model->plist(ip) + 1]; ++iCol)
-                    sx.at(row[iCol], ip) -= data_ptr[iCol];
+                    sx.at(static_cast<int>(row[iCol]), ip) -= data_ptr[iCol];
             }
 
             solveLinearSystem(sx[ip]);
